controller

+++ Controllers/HomeController.cs
@@ -8,50 +8,53 @@ namespace ApiNet6.Controllers
     public class HomeController : ControllerBase
     {
         [HttpGet("/")]
-        public List <TodoModel> Get([FromServices] AppDbContext context) 
-        {
-            return context.Todos.ToList();
-        }
+        public IActionResult Get([FromServices] AppDbContext context)
+            => Ok(context.Todos.ToList());
 
         [HttpGet("/{id:int}")]
-        [Route("/")]
-        public TodoModel GetById([FromRoute] int id, [FromServices] AppDbContext context)
+        public IActionResult GetById([FromRoute] int id, [FromServices] AppDbContext context)
         {
-            return context.Todos.FirstOrDefault(x => x.Id == id);
+            var todos = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (todos == null)
+                return NotFound();
+
+            return Ok(todos);
         }
 
         [HttpPost("/")]
-        public TodoModel Post([FromBody] TodoModel todo, [FromServices] AppDbContext context)
+        public IActionResult Post([FromBody] TodoModel todo, [FromServices] AppDbContext context)
         {
             context.Todos.Add(todo);
             context.SaveChanges();
 
-            return todo;
+            return Created($"/{todo.Id}", todo);
         }
 
-        [HttpPut("/")]
-        public TodoModel Put([FromRoute] int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
+        [HttpPut("/{id:int}")]
+        public IActionResult Put([FromRoute] int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
             if (model == null)
-                return todo;
+                return NotFound();
 
             model.Title = todo.Title;
             model.Done = todo.Done;
 
             context.Todos.Update(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
 
-        [HttpDelete("/")]
-        public TodoModel Delete([FromRoute] int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
+        [HttpDelete("/{id:int}")]
+        public IActionResult Delete([FromRoute] int id, [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (model == null)
+                return NotFound();
 
             context.Todos.Remove(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
     }
 }
